Fix legacy log cleanup potentially deleteing the log buffer state
Add a script that will just delete all old legacy JSON files

+++ unmanic/libs/logs.py
@@ -435,6 +435,8 @@ class ForwardLogHandler(logging.Handler):
             for legacy_name in os.listdir(self.buffer_path):
                 if not legacy_name.endswith(".json"):
                     continue
+                if legacy_name == self.STATE_FILENAME:
+                    continue
                 legacy_path = os.path.join(self.buffer_path, legacy_name)
                 try:
                     os.remove(legacy_path)
